backend/pyproject.toml
[project]
name = "hyphagraph-backend"
version = "0.1.0"
description = "HyphaGraph backend – hypergraph-based evidence engine"
requires-python = ">=3.12"

dependencies = [
    # --- Web / API ---
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",

    # --- Validation ---
    "pydantic>=2.6",
    "pydantic-settings>=2.2",

    # --- Database ---
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",

    # --- Migrations ---
    "alembic>=1.13",

    # --- Utils ---
    "python-dotenv>=1.0",

    # --- Auth (custom JWT, NOT FastAPI Users) ---
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",

    # --- Rate Limiting ---
    "slowapi>=0.1.9",

    # --- Email ---
    "aiosmtplib>=3.0",  # Async SMTP client
    "email-validator>=2.1",  # Email validation
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.8",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.8",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"